+++ fxmanifest.lua
@@ -3,7 +3,7 @@ lua54 'yes'
 game 'gta5'
 
 description 'Police Vehicle Armor Script'
-version '1.1.0'
+version '1.2.0'
 author 'Kanayu_u'
 
 
@@ -15,9 +15,18 @@ shared_scripts {
 }
 
 client_scripts {
+    'notify.lua',
     'client.lua'
 }
 
 server_scripts {
     'server.lua'
+}
+
+ui_page 'html/notify.html'
+
+files {
+    'html/notify.html',
+    'html/notify.css',
+    'html/notify.js',
 }